+++ functions.php
@@ -3,6 +3,7 @@
 require_once get_template_directory() . '/inc/styles-and-scripts.php';
 require_once get_template_directory() . '/inc/shortcodes.php';
 require_once get_template_directory() . '/inc/jetpack.php';
+require_once get_template_directory() . '/inc/authentication.php';
 require_once get_template_directory() . '/inc/search.php';
 require_once get_template_directory() . '/inc/posts.php';
 require_once get_template_directory() . '/inc/volumes.php';